feat(cdk): rename partial_message package
