+++ airbyte-cdk/java/airbyte-cdk/core/src/main/kotlin/io/airbyte/cdk/integrations/destination/async/model/PartialAirbyteStreamState.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.cdk.integrations.destination.async.partial_messages
+package io.airbyte.cdk.integrations.destination.async.model
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import io.airbyte.protocol.models.v0.StreamDescriptor
