+++ airbyte-cdk/java/airbyte-cdk/db-destinations/src/test/kotlin/io/airbyte/cdk/integrations/destination/jdbc/TestJdbcSqlOperations.kt
@@ -4,7 +4,7 @@
 package io.airbyte.cdk.integrations.destination.jdbc
 
 import io.airbyte.cdk.db.jdbc.JdbcDatabase
-import io.airbyte.cdk.integrations.destination.async.partial_messages.PartialAirbyteMessage
+import io.airbyte.cdk.integrations.destination.async.model.PartialAirbyteMessage
 import java.sql.SQLException
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
